Update patch set 2
Patch Set 2: Code-Review+2

A test would be nice.

Also the way zuul is intended to be used is as a gatekeeper. That means it should be the only one doing 'submit' (aka merges). However I realise that may not always be the case so having this to protect us is useful. I do wonder if it should be a config option or something to avoid the extra query though.

Patch-set: 2
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2