Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Just a small nit inline. Otherwise the code looks good to me.

However we've tentatively said no more features in v2 (aka master) until v3 is done. I'm not sure if this is something we want to make an exception for or not, so will defer to James or Jeremy.

Patch-set: 1
Label: Code-Review=-1