Enhance mutex to counting semaphore
The mutex in zuul is great but is limited to run one job at the same
time. Some use cases like using a limited number floating licenses in
jobs cannot be handled with this. Thus this enhances to mutex
functionality to a counting semaphore (which defaults to 1) while
keeping backwards compatibility with the mutex from configuration
point of view.

Change-Id: Ida589e49bc6694f4ccc4c586e0d43b391b8c3ae4