Cloner: use zuul_url always when project set
In the case of a post-merge pipeline (ie, one that
responds to ref-updated events) this increases the accuracy of
a multi-repo checkout by consulting the zuul merger when
cloning and updating the project in question.  All other projects
are updated from the upstream git url.  Previously all repos would
be cloned and updated from the upstream url.  That is normally
fine, but in some cases they may actually lag behind, but the
merger will have the most current copy of the repo for the project
in question.

This is safe because we now ensure that mergers fully update the
branches and tags of their local copies.

Change-Id: If22d8cfbe591afe4f5516da1eb3b0be98b2de874