Update patch set 3
Patch Set 3: Code-Review-1

I don't think the approach of storing the number of retries in a dict on the launcher keyed by job name will work -- we might have many builds with the same job name running simultaneously.  To make that key fully unique, we would have to add in many other pieces of information.  Instead, why don't we keep the counter on the BuildSet.  When a build is removed from the buildset, increase the buildset's counter for that build's job (a job *is* unique within a buildset).  When a build is added to a buildset, pass the current value of the counter for that build's job to the build.  Then, in the launcher, you could consult the build counter and return JOB_LIMIT (or, perhaps "RETRY_LIMIT"?) if it exceeds the retry max.

Patch-set: 3
Label: Code-Review=-1