Refactor test_zuul_refs and FakeBuild.hasChanges
In re-enabling test_zuul_refs and refactoring it to use
FakeBuild.hasChanges, a weakness was discovered in hasChanges where it
would not check the repositories of all the changes one is looking for.

After fixing that, the test passes and others should be able to be
refactored in the same way.

Change-Id: Iaf647412d2518c079c8b42ed670919f4e8ca0b63
Story: 2000777
Story: 2000773