Refactor skip-if tests to use irrelevant-files
Skip-if has been removed in favor of irrelevant-files and the branch
matchers. This refactors the tests that used to be used for skip-if to
cover irrelevant-files.

There may still be need to cover the other types of matchers.

The 'test_parse_skip_if' has just been removed as it wasn't really
covering anything not also covered by the functional tests.

Change-Id: I3a4080426c4dc680ed656e18f26597e3d1b7d759
Story: 2000773
Task: 3378