Update patch set 1
Patch Set 1:

Based on change https://review.openstack.org/#/c/67858 and rebased onto current master.

Concluding the outstanding objections in the former change:

https://review.openstack.org/#/c/67858/8/doc/source/zuul.rst@84
Objection that the SSH might not be suitable for all cases, thus the HTTP should not be considered deprecated. Documentation is updated to describe the SSH as an alternative to access gerrit's git.

Another objection was that waiting for the ref to replicate to local mirror is a hack itself and should be removed altogether. However there was no progress in this so far and I think it is better to get the current behaviour working in different setup where HTTP access is not an option.

Patch-set: 1
Reviewer: Gerrit User 15215 <15215@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0