Move and re-enable test_build_configuration_*
This renames the TestSchedulerMergeModes -> TestSchedulerMerges
and moves some other tests that deal with merge behavior to there:

  * test_build_configuration_conflict -> test_merge_conflict
  * test_build_configuration_branch -> test_merge_branch
  * test_build_configuration_multi_branch -> test_multi_branch

Also, test_merge_conflict discovered a bug in the launcher merging
that caused conflicting merges not to abort the job. It's fixed here
by failing the job if the merger is unable to merge.

Story: 2000773
Task: 3424

Change-Id: I213cd8036addd11ee9d227f94a2ae21b9a168413