Fix conflict detection and reporting for in-repo configs
This fixes up some bugs with conflicting merges in the merger
and the launcher.

It also updates processing and reporting logic in pipeline manager for
handling conflicts with in-repo configs.  Previously, it assumed a layout
was already parsed and would only report conflicts when that layout had
configured jobs for the change.  We now need to detect and report on
these with or without a layout.

Change-Id: I38b4fac4868a823b0de2f6c505744a927c577fad