Update patch set 4
Patch Set 4:

(1 comment)

So why here do we really need to still expose a connection to the core parts of zuul?

Conceptually to me with this change a driver is the interface to a new external source, and a connection is something that is shared (depending on implementation) behind the scenes between the reporter, source and triggers that are fetched from the driver.

Is it legacy or design that zuul should still have direct access to this object? It's strange to me you can save all those connections and use the connection to find the associated triggers etc rather than use the driver interface. 

In part this looks like it could be done to preserve the current configuration model, but do you still see the need for connections to be maintained from the core like this?

Patch-set: 4