Update patch set 4
Patch Set 4:

Jamie,

We discussed this in IRC a bit, and the outcome was basically that the biggest obstacle to hiding the connection concept is that since connections are defined in zuul.conf and then referenced in zuul.yaml (in triggers, etc), that when zuul.yaml references a connection on a trigger, zuul needs to be able to identify which driver is associated with that connection.  Having connections as a concept that zuul knows about facilitates that in the current design.

Patch-set: 4