Update patch set 2
Patch Set 2:

Yeah, I use them a lot too, which is why we set this up so that it runs this way with tox, but if you use ttrun, you get debug logs by default -- I personally depend on that when running locally...  We restrict them in tox so that we don't run afoul of the 50mb gate limit.

(It would be a magic unicorn if we could post-process the testr logs to remove logs from *passing* tests and only include the failing ones.  Then we could have full debug in the gate.)

In the mean time, I'm trying to strike a balance to get us logs likely to be able to identify a problem only seen in the gate.

Patch-set: 2