Add github reporter status_url config option
Make it possible to set the github status link to something other than the
zuul-server status url. This allows the github status to link to the log output
for the jobs. If set, this url will be run through the same formatter
as the url_pattern used for comment links. If unset, it will fall back to the
previous behaviour of using the zuul_server status_url.

Since we no longer ignore formatting errors, remove references to build.number
from the tests (number was remove from Build objects in change
I75525a2b48eb2761d599c039ba3084f09609dfbe)

Change-Id: I5f2ff35bb38426ddcd7f65798b4f114256a54847