Fix infinite loop when Gerrit close connection
Gerrit may close SSH connection remotely (ie: timed out) on stream
events. When it happens, select is continuously sending POLLIN events
while stdout is still empty. The connection is then stuck in a loop.
I fix it by detecting that the input line is empty, and then raising
a standard Exception, that is catched, and restarts the ssh connection

Change-Id: I7cb43f688f00e985908676fdd14ba094e88f5b86