Copy nodeset when making node requests
Nodesets are collections of nodes, but they, and the Node objects
they contain, have a dual role representing both the configuration
of a job and the actual nodes returned from nodepool.  So that
multiple jobs sharing the same nodeset configuration don't end up
sharing the same actual nodes, create a copy of a job's nodeset
for use in interacting with nodepool/zk.

Change-Id: I83f503ba22fc3f92b8c90b15ccfb6b07dc3c4709