Re-enable test_build_configuration_branch
This test was relying on the implicit shared queue of org/project. So
we're moving it to org/project2, which has the 'integrated' queue to
make sure all the changes are enqueued in order. We can't really rely
on the refs anymore, as the merger has changed, nor can we rely on the
local state of the git repository because of the way FakeGerritConnection
effectively hard-resets the repository. So we just check to make sure
all 3 changes are in the last merge build.

Change-Id: I9d28e81c3dc7c0d5af5753a291628787ba21c22d
Story: 2000773
Task: 3443