Add support for requiring github pr head status
This hooks into what exists for gerrit, a requirement on status. However
it slightly modifies the model so that the change status could be a list
or a str, as GitHub commits can have more than one status.

When creating a github pull request object, the statuses of the head
commit are fetched, and the latest status of each user and context set
is appended into a list of statuses. The string is 'user:context:state',
where state can be success, error, or failure. Context is freeform, and
the user is the login of the entity that set the status.

Tests have been added to validate that requirements on github based
pipelines are honored.

Tests that would handle trigger requirements are commented out, as we do
not yet support triggering on status, so we can't test filtering that to
only trigger on status from particular user:context:state.

Change-Id: I45abbd6cbddd36b8491bdf9bb8d545216537ad2f
Signed-off-by: Jesse Keating <omgjlk@us.ibm.com>