Put Zuul vars in an ansible vars file
This adds a vars file for use by ansible playbooks with all vars
scoped under 'zuul'.  The existing environment variables will be
moved into this section in a later change.  Currently, the only
supplied variable is 'uuid' for use in a test.

Also, add a test-specific vars entry (zuul._test) so that we can
pass information such as the test chroot into playbooks used in
our tests.  This is used by a test to set a flag file inside of
the test chroot to verify that the ansible playbook in a job
actually ran and did something.

Change-Id: Ie5d950d051accad4ec9dc90a9e1b01b3095a1e5c
Co-Authored-By: Monty Taylor <mordred@inaugust.com>