Handle multiple playbook extensions
Detecting whether to look for a .yaml or .yml file for a playbook,
and detecting whether that playbook exists would ideally be done
in the scheduler.  However, doing so involves passing and storing
quite a bit of extra data (file lists for each project-branch
combination), and might put a crimp in making playbook specification
more sophisticated later.  So for now, let's do it in the launcher
where it's easy for us to test for the presence of files right
before we run a job.  Even though that means we won't detect errors
until later, for many changes this will still be self testing and
should prevent many config errors from landing.

If need be, we can do the extra work to move it into the scheduler
later.

Change-Id: I1ad2eb4a5d0ff08fbd2070f55e352633dd6de81b