Fix nodeset return on skipped jobs
Change Ideab6eb5794a01d5c2b70cb87d02d61bb3d41cce altered the
node request algorithm so that nodes are not requested unless
the job is ready to run.  Therefore, this stanza which returned
nodesets for skipped jobs would always fail (since skipped jobs
are jobs which were waiting for their parent to be ready to run
and therefore had no nodeset).

This likely contributes to some test instability as it would cause
the main loop to exit with an exception.  While it would restart
and recover, the test may decide it was finished before that could
happen.

Change-Id: I189bd59fa87460a1097fd6c6ebc04f29a89674cd