Add reporter for Federated Message Bus (fedmsg)
Fedmsg is a bus that is heavily used by the Fedora project. By adding
support for fedmsg, we can start to do cool things with the Fedora
projects infrastructure.  As an example, fedmsg-notify could be used
to provide desktop notifications.

Fedmsg also has a plugin system for brokers, which allows us to add
support for mqtt.

Change-Id: Id0fa74bd81c3f809a2ec4fb03e960e2e2691c675
Signed-off-by: Paul Belanger <pabelanger@redhat.com>