Tests: store debug logs on error
Rather than using the FakeLogger fixture, which always attaches
the log stream as a detail, use standard loggers that output to
a stringio, and then, only if a test fails, attach the full
log as a detail.

This allows us to report full debug-level logs for failing tests
in the gate (which normally has a limit on how large subunit files
can be).

Switch to the gentle timeout.  This has caused problems in the past,
but perhaps we're better about unclean shutdown now.  Give it another
shot since non-gentle timeouts are so hard to debug.

Also, increase the test timeout.  test_dependent_behind_dequeue is
running right at 30 seconds now.

Change-Id: I9e6509b7b69838d29582b040ef22f1d66010d45e