Update patch set 1
Patch Set 1: Code-Review-1

This test passes, but I think the time database needs further updates to support multi-tenancy and zuulv3 features.

First, the database should be tenant-scoped.  That probably means adding a path component to the files on disk corresponding to the tenant name, and also including the tenant when interacting with the database.

Second, the current database is based on the idea that job names uniquely identify what they run, which is no longer true in Zuulv3.  So in addition to the tenant, we need to scope the entries further to include project and branch information.

Put another way, in zuulv2, "gate-nova-python27" was unique enough for our purposes*.  In zuulv3, "python27" is not unique enough.  We need to specify "python27" for the "master" branch of "nova" in the "openstack" tenant.

* Even in zuulv2, that job could run on multiple branches, so we should have included branch information in order to get the most accurate time.  But we were being bug-compatible with our use of jenkins in this.  Since we're adding all the other specifiers, why don't we go ahead and add the branch here too.

Patch-set: 1
Label: Code-Review=-1