Inherit playbooks as expected
An earlier change dealt with inheritance for pre and post playbooks;
they are nested so that parent job pre and post playbooks run first
and last respectively.

As for the actual playbook, since it's implied by the job name, it's
not clear whether it should be overidden or not.  We could drop that
and say that if you specify a 'run' attribute, it means you want to
set the playbook for a job, but if you omit it, you want to use the
parent's playbook.

However, we could keep the implied playbook behavior by making the
'run' attribute a list and adding a playbook context to the list each
time a job is inherited.  Then the launcher can walk the list in order
and the first playbook it finds, it runs.

This is what is implemented here.

The logic for inheriting playbooks only for "real" inheritance rather
than job variants is extended to pre and post playbooks as well, since
they should follow the same logic.

Change-Id: Iaf6d661a7bd0085e55bc301f83fe158fd0a70166
TODO: all of these conditions need tests.