Re-enable test_idle
This is the second test that wants to switch between two sources. This
time it wants to do that 3 times, so we just reset hard each time.

To make this simpler, we made addCommitToRepo return the commit from
before this one. We also refactor the update from one config fixture to
another into a generic method that should be useful a few more times.

Change-Id: I409a86b3c92cc8f7984967f50ef0db50bf37ea6b
Story: 2000773
Task: 3473