Update patch set 11
Patch Set 11: Code-Review+1

Giving this a +1, but as I mentioned on IRC, it may be more maintainable if we use a plugin or strategy to maintain a whitelist or blacklist of modules that should not be used. I'll be easier to audit at each ansible release, easier to communicate to consumers, etc..  I'm just not sure if it's possible.  Perhaps a thing for future enhancement.

Patch-set: 11
Reviewer: Gerrit User 6589 <6589@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1