Update patch set 3
Patch Set 3:

It's entirely within the gerrit driver, just in two different classes.  If Python had a protected attribute, this would use it.  I think the tail is appropriate here because it's a reminder to us that it's not part of the connection api, it's internal to the gerrit driver.

Patch-set: 3
Label: Verified=0