Correct playbook requirement logic
First, it was a little hard to follow what was going on here because
of the main -> required variable change.  So let's call them both
'required'.

Second, it sure does look like we were saying that we required the
first instance of the main playbook to be present, but not the
pre and post playbooks.  The opposite is true -- every instance
of pre and post playbooks is required, but we only require that
one of the main playbooks exists.  Reverse the logic.

Change-Id: Ibb1ffb725e09b65bcfaac4878ae3d8284864f5fb