Add optional basic auth for merge check
When Zuul polls Gerrit to see if a change has merged, it does so
without authentication. If Gerrit is locked down to require
authentication before permitting read access, this fails after polling
for five minutes. This patch adds optional basic auth to that call,
which is enabled by defining a password for the user Zuul uses to
communicate with Gerrit.

Change-Id: If335dbbcfe4b255af401e53c25139a407ee2e16f