Change 'secure' to 'trusted'
Conceptually, we're trying to express whether we trust the authors of a
job or not, and whether or not we trust the job to be able to have
access to secrets or request plugin execution. Reading and writing
narrative text using the word secure for that starts to hurt the head
sometimes. Switch to trusted.

Change-Id: Ic6a9fe7406f808f965a0ed5ef099fdea92f52c25