Add support for 'repos' job configuration
If a job configuration gives a list of repos, add them to the list of
projects to update on the slave.

Test using a mock openstack dsvm job which should clone both nova and
keystone. Put this new mock job in the check pipeline rather than the
gate pipeline to keep the build history small, and assert that both the
launcher and the worker have cloned the project that did not trigger
the job.

Change-Id: I3ccf8713906d65cbd27929548499e81f798cea82