Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I think this is running into something I haven't quite finished or articulated well in the launcher.  The idea is that the launcher has an internal merger which it keeps up to date with changes.  The code you have here will make sure that it correctly updates the related projects.

But when jobs are run, they actually each get their own copies of all the repos in a new git root inside of their temporary JobDir.  Ideally we will clone the jobdir git repos from the merger's main repos -- essentially using it as a cache.   We haven't implemented that bit right now, so currently we needlessly make entirely new clones from the source.  But that shouldn't be hard to fix.

So we should build on this change and have it check out the additional repos into the jobdir.  The test is a little misleading right now since it looks at the launcher's main merger git root.

Patch-set: 2
Label: Code-Review=-1