Re-enable test_dependent_behind_dequeue
This test was already refactored for v3, but it was causing occasional
timeouts. This is a very merge-heavy test, and as such, sometimes runs
out of time when contending for resources. So we'll bump the overall
hard-timeout by 30 seconds which should prevent us from timing out at
that level, and also raise the soft wait timeout to 90 seconds for just
this test.

Change-Id: I74e54d8a7d9b813772c498ccb87fb6644b533842