A couple small test improvements
If the build history assertion fails, log the details of the historic
build.

Make the use of fixtures.TempDir optional, and if we are not using the
fixtures (which will be automatically deleted after the test run), use
tempfile.mkdtemp so that every ZuulTestCase object gets a unique
directory under the ZUUL_TEST_ROOT path to avoid collisions between
concurrent tests. Additionally, if we are not using fixtures, also set
LaunchServer.keep_jobdir = True.

Change-Id: Ie9a46d0a4b5b075ca2776bbd4667117aebc1a9df