Clone jobdir git repos from the launcher's cache
One of the main reasons we made the update thread in the launcher
was so that we could keep current copies of all the git repos
needed by jobs in the launcher, and then clone from there to the
jobdir so that we generally clone from cache.  We weren't actually
doing that, but instead were cloning from the source for each job.

This changes the jobdir merger to clone from the cached repos.

Change-Id: I2be41424c26028068671aecec2520a4a6ad7ae66