Report layout config errors for config repos
If a change with a syntax error is proposed  to a config-repo, the
error detection and reporting recently added will not activate
because config-repo changes are not dynamically evaluated.  However,
as soon as that change landed, the configuration would attempt to
reload and fail due to the error, potentially putting the operator
in a catch-22 that would require external intervention.

To avoid that, evaluate and construct a dynamic layout *including*
any changes to config-repos for the purpose of detecting configuration
errors.  Then discard that layout and construct a new one *excluding*
config-repos (as we do now) for actual use when running jobs.

Change-Id: I803cb05274e79e91477c202b754a1640391e63a6