Update patch set 1
Patch Set 1: Workflow-1

This is a port of https://review.openstack.org/231124 to the v3 branch.  It generally follows the original, except for necessary changes due to Zuul v3 restructuring.

A significant difference is in the config file syntax: the configuration file job listing format is flattened in all cases due to the presence of job variants in v3.  Essentially, it's not comprehensible to have a YAML tree shape express the dependencies when any job may appear more than once (possibly with a disjoint set of dependencies); attempts to do so end up with fairly major duplication and/or would be very difficult to read.  Instead, jobs list what their dependencies are as any other attribute.

Additionally, due to Zuul v3's late binding of which jobs (and variants) to run, the graph is not constructed until right before jobs are launched, and therefore we are unable to provide configuration-time validation of the graph.

Note, this does not currently pass most tests as nearly all the test configuration files will need
to be mechanically updated to use the new format.

Patch-set: 1
Label: Workflow=-1