Perform pre-launch merge checks
Split prepareLayout into prepareItem and prepareJobs so that we
can schedule merges and react to merge failures before the item
enters the actionable window, but delay layout preparation and
job freezing until the item is ready to start running jobs.

Move prepareItem higher in the loop so that it runs before we check
if the item is actionable. This allows us to schedule a new when
item.active changes to True. This also means we run prepareItem
whether or not we have failing dependent items.

In order to test that a conflict merging in the gate causes merge
conflicts for queued changes in check, expand the release() function
to match jobs base on any build attribute, not just build.name.

If the executor encounters a merger failure, immediately report
that result.

Change-Id: I1c9db6993994bf8e841ecd8554c37a3ec0afc798
Co-Authored-By: Adam Gandelman <adamg@ubuntu.com>
Story: 2000773
Task: 3468