Perform pre-launch merge checks
Split prepareLayout into prepareItem and prepareJobs so that we
can schedule merges and react to merge failures before the item
enters the actionable window, but delay layout preparation and
job freezing until the item is ready to start running jobs.

Also, if the executor encounters a merger failure, immediately report
that result.

Change-Id: I1c9db6993994bf8e841ecd8554c37a3ec0afc798
Co-Authored-By: Adam Gandelman <adamg@ubuntu.com>
Story: 2000773
Task: 3468