Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

We have been leaving doc updates until things settle down more, though this is unlikely to change much, so it probably wouldn't be a waste of time to go ahead and add it.  I'm okay either way.

I left a -1 because this will add an extra space to the ansible inventory file.  I'm assuming that was an accident, but if was intentional and the artistic vision is to have two spaces after the node name, I'll change my vote.  :)

Patch-set: 1
Label: Code-Review=-1