Add support for job allowed-projects
In order to allow users to create jobs which are only available to
run on certain projects, add an 'allowed-projects' attribute to
jobs.  This is especially useful for jobs with secrets.  For example,
consider a job which performs external API testing with credentials.
The user defining that job may well want to restrict its use to a
single project.

Change-Id: I4457842ea293baf20b83f7b8b86fba2b7d26d2be