Update patch set 3
Patch Set 3: Code-Review-1

I was suggestion a partial revert of the code which removed the pre-launch merge checking for two main reasons: first, we need to perform pre-launch merges some of the time any way (when a .zuul.yaml file changes) so therefore, much of the functionality I was hoping to remove must remain.  Second, by the time the executor has the job in hand, nodes have already been allocated and marked as used, so it's quite wasteful to perform the check there.

So I think the solution should principally involve restoring the merge check that happens in the pipeline manager before a job is launched.

Patch-set: 3
Label: Code-Review=-1