Update patch set 8
Patch Set 8: Code-Review-1

(4 comments)

Unrelated to the comments inline, I'm confused by the general structure of the playbooks/ content. Some of the playbooks simply call a single role. Some of the playbooks have actual tasks in them and do not call a role. None of the roles depend on each other. What really is the point of having single role plays, of which the roles have minimal tasks? As a developer adding new stuff in, how would I know to choose a playbook with direct content vs introducing a new role?

Patch-set: 8
Label: Code-Review=-1