Update patch set 14
Patch Set 14: Code-Review+2

I was going to write that we should call this role "python sdist" or something, however, I realize that this is just running the openstack run-tarball script, so it is not going to end up in the stdlib and is appropriately named.  However, if we replace it with an ansible re-implementation (and make it somewhat more universal), it would be appropriate to include it in the stdlib and give the replacement a better name.

Patch-set: 14
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2