Move dependency cycle detection into pipelines
This changes the gerrit connection to tolerate cyclic patch
dependencies. This is a first step towards support of cyclic patch
dependencies where this will need to be handled in the pipelines
themselves. For now the pipelines are changed to detect the cycles by
themselves allowing to enhance them for cyclic dependencies later on.

Change-Id: I26e37b3fe66f46250a1baea1b803fd82d07145d7