Update patch set 6
Patch Set 6: Code-Review-1

(2 comments)

In addition to the in-line comments, it's not clear how the secret block added to the layout gets tested. Seems the same data is added to a on-the-fly layout in one test file. I'm probably missing something. The -1 is for the name of "data" though.

Patch-set: 6
Label: Code-Review=-1