Fix constructor arguments to source
Instatiations of sources all passed (driver, connection) to the
constructor, which was expecting (config, connection).  The
config option is currently ignored (sources currently have no
additional configuration), which is why we didn't notice the
discrepancy.

Update the constructor to (driver, connection, optional config)
to match the rest of the driver-related classes.

Change-Id: Ibc878b51b81950559d39b00b1591864c7661fe7c