Add source to project and remove unused tenant attrs
This adds to the project class a reference to the source which
created it.  This forms a cycle with connection->project->source,
but we expect these objects to all have similar lifecycles (they
are likely to persist through the lifetime of a tenant, but may
be replaced during a full reconfiguration).

This lets us drop the (source, project) tuples we were carrying
around on the Tenant class which simplifies it.

Change-Id: I5e72255815131493c516a84463dcfd7100f4e827
Story: 2000953