Add support for bwrap
We add the ability to set an arbitrary insecure prefix/suffix, and then
provide one example of how one can do that.

This will be the minimum "batteries included" bubblwrap script. It does
not do any MAC configuration, since these vary by system. Operators
may wish to wrap this in a MAC wrapper. It also does not setup seccomp,
but it should at some point once we learn what is necessary for running
playbooks.

The test sets up bwrap-executor.sh as the wrapper for a positive
execution test. But it lacks coverage for negative (breakout of Ansible
is contained) and also lacks coverage for suffixes, which may need to
be dropped but were added here for completeness.

Change-Id: I84dd7c8cc55d2110b58609784007ffda0d135716