Update patch set 2
Patch Set 2:

The follow up change (I26e37b3fe66f46250a1baea1b803fd82d07145d7) breaks without this patch. Though this is unrelated to it so I put it into a separate patch. Reason is that when building lists or dicts with the change number of the event (gives str) and change numbers from querying the source (gives unicode), existence checks for these number in lists behave incorrect as the same number is treaded different for str and unicode. If needed I'll try to find a test case (not sure if current code triggers this behaviour though as I discovered this misbehaviour during debugging why the tests of the follow up patch fail).

If preferred I can also squash this with the follow up as there it's needed but I thought making a separate patch for that is clearer.

Patch-set: 2
Reviewer: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0