Create wrapper script for executor
This will be the minimum "batteries included" bubblwrap script. It does
not do any MAC configuration, since these vary by system. Operators
may wish to wrap this in a MAC wrapper. It also does not setup seccomp,
but it should at some point once we learn what is necessary for running
playbooks.

Change-Id: I52bb15d744f1ca7cba96be54a91d29b81b47b6b7