Update patch set 2
Patch Set 2:

Sorry, I was mistaken. Thought this was the cause of the test failures in the follow-up. Just tested the follow-up again without this and it still worked.

But I still think it's cleaner to use the correct type here instead of relying on magic type conversions. But I don't mind abandoning this if that's whished.

Patch-set: 2