[WIP] Integration: Get static nodes from nodepool
This is the first part in an attempt to provide a multinode integration
test in zuulv3. As discussed in the IRC log linked below, the approach
we have decided on is to get some static nodes from zuul's fake
nodepool and use them for running a full multinode integration test.

This review is essentially a slight modification of the current unit
tests for nodepool, and creates an integration directory for use in
subsequent reviews.

Note: I am open to approaching this in a different way, but want to
throw up small commits to make the review process digestable and make
sure this is getting started correctly.

Relevant IRC log:
http://eavesdrop.openstack.org/irclogs/%23zuul/%23zuul.2017-03-29.log.html#t2017-03-29T15:49:37

Storyboard link:
https://storyboard.openstack.org/#!/story/2000878

Change-Id: Ib41b7089940525de5271f5d0f237b6fea56dc606