Update patch set 8
Patch Set 8:

Looks like a few of the tests need addressing.

I like this rename, I think "config-project" makes a lot more sense as it highlights the key difference. However, I'm not sure "untrusted-project" is as true to the key point for those projects.

To me the "untrusted-projects" key feature is that their tests are done in-tree. They are the real projects that we are testing, and anything else is CI/config related.

My suggestion would be to rename "untrusted-projects" to just "projects" again.

I feel like the configuration difference is explained by the "config-project"s, in that access to various config options/syntax is reserved for config-projects. So I don't think we need to further emphasise untrusted-projects. It's generally accepted that build environments are somewhat locked down, so I think that's also covered.

(Some more suggestions: "build-projects", "tested-projects", "ci-projects")

Patch-set: 8
Label: Verified=0