Don't report start on unmanaged projects
When using a check pipeline with enabled start reporting zuul reports
start on all projects of the gerrit instance regardless if they are
part of the pipeline or not. After that zuul notices that there are no
jobs to do.

The fix is to skip creating a dynamic change queue when adding a new
change.

This also adds the getChangeQueue function to the PipelineManager base
class as it's called there and implemented in both subclasses.

Change-Id: Idd754f3e42157d50012f49ef89cfb05f566efd7e