Handle dict extras in Voluptuous Schema
Not sure if schema requirements changed but it is no longer valid to
use:

  Schema({}, extra=True)

The docs say that should be avoided and instead you should use:

  Schema(dict)

To indicate a dict that can have any key: value pairs. Additionally
extra=True appears to be replaced with extra=ALLOW_EXTRA so update that
in cases where we have some subset of desired keys.

Change-Id: I3dc966dae534416f2b601ec8d299e1d0e3651e7b