Add a finger protocol log streamer
This will be started along side the the executor process, similar
to how the scheduler and gearman server are linked.

This will require starting the process as the root user so that we
can grab the finger port properly. The process listening on the
finger port will drop its privileges to the designated user after
grabbing the socket. The executor will also drop its privileges
to the same user after starting the log streamer.

Change-Id: Ib52585cafbd073ccdb7f87432888ce15c7a66f67