Don't run apsched thread when no jobs need scheduling
We leak unneeded apsched threads when we reconfigure to stop running
jobs on a timer. This is mostly fine because as soon as you add new jobs
to the timer that thread will be used again, but in the interest of
tidyness and log sanity in testing make sure we stop the apsched thread
when we don't need it.

Note that you can't restart an apscheduler after shutting it down.
Therefore we create a new apscheduler object when we restart the need
for timer based jobs.

Change-Id: I5aa081451a7598b3ed67e3dbc5280290e1db7cff