Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I think inline comment is source of pep8 failure.

I like this approach, but think we should work towards making "simple_layout" the default and invert the decorator so that complicated tests are the special ones. I realize this is transitional hence starting here with this setup. But would be good if we could make simple the default so that its easy to write test cases in the normal case.

Patch-set: 1
Label: Code-Review=-1