Remove org_unknown git repo from single-client test config
This was used by two tests to simulate a repo that Zuul doesn't know about.
Using simple_layout is not appropriate here because that would add it to
Zuul's configuration.  The tests aren't structurally a great fit for moving
to a new class with a dedicated config directory, so instead, just make
init_repo work for creating the unknown repo in the fake gerrit without
including it in Zuul's configuration (which is closer to what the test did
in v2).

Change-Id: I60d8897152e3fcfe48f85554604815b3ef86c5b3