Move semaphore tests to their own class
Create a dedicated config directory for the semaphore tests and
remove them from the single-tenant configuration.

Create a simplified form of commitLayoutUpdate which accepts a
path to a replacement zuul.yaml and commits it to the specified
config repository to aid in reconfiguration tests.  The existing
similar methods rely on an entire shadow git repository which
requires additional git filesystem operations in tests.

Change-Id: I0f8e99b6ad262ece5a5649a480e0393872761903