Join watchdog thread so it doesn't leak in tests
This thread was seen to leak in tests because the stop method didn't
wait for the thread to join and that thread only checked its running
status (set by stop to false) every 10 seconds. Reduce the sleep to 1
second and join on the thread so that stopping the watchdog means the
watchdog thread has stopped.

Note that this change increases the timeout in the timeout test job to 3
seconds from 1 second because the new one second sleep in the watchdog
is close enough to the old 1 second timeout failue to get false positive
timeouts. Previously it wasn't an issue because any playbooks run
completed in under the 10 second sleep in the watchdog.

Change-Id: If3cf86c4af0b2fbcaf51a233c766c75749ca1d1d