Join watchdog thread so it doesn't leak in tests
This thread was seen to leak in tests because the stop method didn't
wait for the thread to join and that thread only checked its running
status (set by stop to false) every 10 seconds. Reduce the sleep to 1
second and join on the thread so that stopping the watchdog means the
watchdog thread has stopped.

Change-Id: If3cf86c4af0b2fbcaf51a233c766c75749ca1d1d