Update patch set 2
Patch Set 2: Code-Review-1

We're going to have a lot (about 100 in openstack-infra?) watchdog threads running on an executor, so having the watchdog poll interval be so short will make a significant difference (especially considering the python GIL).

How about we plumb an internal parameter from the tests through the executor to the watchdog to set the poll interval?  It should not be user-configurable, it should just be hard-coded to 10 unless the tests override that to be 1.

However, even that will cause a 0-10 second delay after finishing every job in production.  So maybe we can also make the join optional (only do it in tests), or, as we discussed in IRC, leave the code as-is (10s, no join), but whitelist the watchdog thread for the end-of-test assertions.

Patch-set: 2
Label: Code-Review=-1