remove getSourceByHostname
This erroneously assumed we could only have one source with a
given hostname.  Instead, the project->source mapping in the
tenant config file is key.  Therefore, to find a source, we need
a fully qualified project name, and must use the tenant project
lookup to find the source for that project.

Change-Id: I7ab9c4222014786c14eea301a5a419b8318b9461