Use routes for URL handling in webapp
The custom regular expressions were a bit difficult to manage and there
are plenty of libraries that can do that for us. Use Routes which
handles purely URL matching rather than a full web framework. This will
at least let us rationalize the list of URLs as the REST api becomes
more established.

Change-Id: Ia9eb45a51d5c1638529b461426e42f8592b95097
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>