Use project canonical hostnames in mergers and executors
Use the canonical hostname within merger operations so that
mergers which manage repos for different connections don't have
conflicts if they have different projects with the same name.

Since the executor uses a merger to manage the src directory in
the jobdir, this causes the checked out version of repositories
to include the canonical hostname.

Change-Id: I24cb909032b343af653b191fd6f553a018c993f3