Use connecction to qualify projects in merger
The previous change fully qualified projects in the merger with
canonical hostnames.  Keep that format on disk (so that the on-disk
representation in the merger matches what we expect in the executor
and also survives changes in connection names), however, switch to
using the connection name to uniquely identify projects in the merger
API.  This lets us drop the URL parameter (which always seemed
unecessary, as the merger can figure that out on its own given a
uniquely identified project).

This simplifies both the API and the JSON data structure passed to
the merger.

The addProject method of the merger is flagged as an internal method
now, as all "public" API methods indirectly call it.

In the executor, after cloning and merging are completed, the 'origin'
remote is removed from the resulting repositories since it may not
be valid for use within a running job.

Change-Id: Iaf8a16eb152fd12448c74f4a7053926c6d2a27de