Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

Generally looks good, but I'm not sure why we're losing the ability to unregister a path.  We're going to need that in the future in the case that a connection is removed from the configuration (or if some driver's configuration is complex enough that it needs to change the routes on reconfig).

Patch-set: 4
Label: Code-Review=-1