Update patch set 6
Patch Set 6:

Wow, it's really unfortunate that routes is so inflexible.  To be honest, I'm losing enthusiasm for using routes.  At a high level, it looks like it's basically like what we have now, except that instead of the fairly standard regular expression syntax, it uses its own syntax, and we lose the ability to remove routes on reconfiguration.

I don't want to fall into a NIH trap, but in this case, we have something that does exactly what we want and we're replacing it with something that does almost what we want.  And in exchange, we still don't get to remove any code.

I'm not opposed to this, but I'd like to be convinced why it's compelling.

Patch-set: 6
Label: Verified=0