Update patch set 6
Patch Set 6:

The one thing I see here that I think is better than what we had before, the ability to easily register redirects, which may be important as things evolve and change? 

It does not appear to save us much code per se, and it brings in a host of new code by way of the library to worry about. Every library ahs a cost, but then again so does every NIH, so yeah, it's a balance. Is there more of a case that can be made for this change?

Patch-set: 6
Reviewer: Gerrit User 6589 <6589@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0