Update patch set 6
Patch Set 6:

Yea, i agree that when we have to do custom handling for the connection routes this unfortunately  becomes less compelling. It's annoying routes doesn't have this feature - but i can see why as it's maintaining a list of re.compiled paths that may be difficult to remove from. 

So with routes or not, what bugs me about the old method and we should try to change is the ambiguity of what that routing table looks like. I still have trouble figuring out the code flow from that old app() method and which routes have tenant_name prefixes and which don't.

The main advantage of routes to me is looking at that one function and being able to determine the entire accepted API. We can do this without routes, it just seemed easier to use something existing.

Long term it would be nice if this was standalone and used gearman/zk for getting that status data, and at which point we could put in a real full web framework.

Patch-set: 6
Reviewer: Gerrit User 7191 <7191@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0