Implement github trigger requirement status
This allows using a connection specific requirement on the status of the
head of a PR. A list of statuses is accepted, just like the pipeline
requirement.

This updates the githubmodel so that there is a common set of filters
that can be used both by the eventfilter and by the ref filter.

Change-Id: Ib91f6527bf1d8ff5fbc6434c8adaca1cd5e1ba6d